Update so sensor can be added to a system after the first initialisation.

+++ eodatadown/eodatadownGEDIsensor.py
@@ -282,7 +282,7 @@ class EODataDownGEDISensor (EODataDownSensor):
                 self.parse_plugins_config(config_data["eodatadown"]["sensor"]["plugins"])
             logger.debug("Found the plugins params")
 
-    def init_sensor_db(self):
+    def init_sensor_db(self, drop_tables=True):
         """
         A function which initialises the database use the db_info_obj passed to __init__.
         Be careful as running this function drops the table if it already exists and therefore
@@ -291,8 +291,9 @@ class EODataDownGEDISensor (EODataDownSensor):
         logger.debug("Creating Database Engine.")
         db_engine = sqlalchemy.create_engine(self.db_info_obj.dbConn)
 
-        logger.debug("Drop system table if within the existing database.")
-        Base.metadata.drop_all(db_engine)
+        if drop_tables:
+            logger.debug("Drop system table if within the existing database.")
+            Base.metadata.drop_all(db_engine)
 
         logger.debug("Creating EDDGEDI Database.")
         Base.metadata.bind = db_engine
