+++ eodatadown/eodatadownsentinel2googsensor.py
@@ -393,7 +393,7 @@ class EODataDownSentinel2GoogSensor (EODataDownSensor):
                 self.parse_plugins_config(config_data["eodatadown"]["sensor"]["plugins"])
             logger.debug("Found the plugins params")
 
-    def init_sensor_db(self):
+    def init_sensor_db(self, drop_tables=True):
         """
         A function which initialises the database use the db_info_obj passed to __init__.
         Be careful as running this function drops the table if it already exists and therefore
@@ -402,8 +402,9 @@ class EODataDownSentinel2GoogSensor (EODataDownSensor):
         logger.debug("Creating Database Engine.")
         db_engine = sqlalchemy.create_engine(self.db_info_obj.dbConn)
 
-        logger.debug("Drop system table if within the existing database.")
-        Base.metadata.drop_all(db_engine)
+        if drop_tables:
+            logger.debug("Drop system table if within the existing database.")
+            Base.metadata.drop_all(db_engine)
 
         logger.debug("Creating Sentinel2GOOG Database.")
         Base.metadata.bind = db_engine
